# ErpApi/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ErpApi",
    "description": "API para cadastramento de FORNECEDORES vinculados a uma EMPRESA. <br/><br/>***ATENÇÃO: utilize (1) para FISICA e (2) para JURIDICA em tipoPessoa.<br><br> Formato para busca por data ==>  yyyy-MM-dd ",
    "version": "v1"
  },
  "paths": {
    "/api/Empresa": {
      "get": {
        "tags": [
          "Empresa"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Empresa"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Empresa"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Empresa/{id}": {
      "get": {
        "tags": [
          "Empresa"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Empresa"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EmpresaDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Empresa"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Fornecedor": {
      "get": {
        "tags": [
          "Fornecedor"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Fornecedor"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Fornecedor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/{id}": {
      "get": {
        "tags": [
          "Fornecedor"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Fornecedor/nome/{nome}": {
      "get": {
        "tags": [
          "Fornecedor"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "nome",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/cpf/{cpf}": {
      "get": {
        "tags": [
          "Fornecedor"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cpf",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/cnpj/{cnpj}": {
      "get": {
        "tags": [
          "Fornecedor"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cnpj",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/dataNascimento/{dataNascimento}": {
      "get": {
        "tags": [
          "Fornecedor"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "dataNascimento",
            "required": true,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Fornecedor/{id}": {
      "put": {
        "tags": [
          "Fornecedor"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Fornecedor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Fornecedor"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "ETipoPessoa": {
      "format": "int32",
      "enum": [
        1,
        2
      ],
      "type": "integer"
    },
    "Fornecedor": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "tipoPessoa": {
          "$ref": "#/definitions/ETipoPessoa"
        },
        "cpf": {
          "type": "string"
        },
        "rg": {
          "type": "string"
        },
        "cnpj": {
          "type": "string"
        },
        "dataeHoraCadastro": {
          "format": "date-time",
          "type": "string"
        },
        "dataNascimento": {
          "format": "date",
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "empresaId": {
          "format": "int32",
          "type": "integer"
        },
        "empresa": {
          "$ref": "#/definitions/Empresa"
        }
      }
    },
    "Empresa": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nomeFantasia": {
          "type": "string"
        },
        "uf": {
          "type": "string"
        },
        "cnpj": {
          "type": "string"
        },
        "fornecedores": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Fornecedor"
          },
          "readOnly": true
        }
      }
    },
    "EmpresaDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nomeFantasia": {
          "type": "string"
        },
        "uf": {
          "type": "string"
        },
        "cnpj": {
          "type": "string"
        }
      }
    }
  }
}